[refactor] move Attack from npc.py to combat.py
